pullback_bot/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
websockets>=12.0
pandas>=2.2.0
aiosqlite>=0.20.0
python-dotenv>=1.0.1
httpx>=0.27.0
numpy>=1.26.0
